all the .buzz files are buzz's versions that need review

+++ Tools/scripts/build_binaries.py
@@ -14,18 +14,28 @@ import os
 import re
 import shutil
 import time
+import string
 import subprocess
 import sys
 import gzip
 
 # local imports
-import generate_manifest
+import generate_manifest, gen_stable
+import build_binaries_history
+
+if sys.version_info[0] < 3:
+    running_python3 = False
+else:
+    running_python3 = True
 
 
 class build_binaries(object):
     def __init__(self, tags):
         self.tags = tags
         self.dirty = False
+        binaries_history_filepath = os.path.join(self.buildlogs_dirpath(),
+                                                 "build_binaries_history.sqlite")
+        self.history = build_binaries_history.BuildBinariesHistory(binaries_history_filepath)
 
     def progress(self, string):
         '''pretty-print progress'''
@@ -74,6 +84,10 @@ class build_binaries(object):
                     # select not available on Windows... probably...
                 time.sleep(0.1)
                 continue
+            if running_python3:
+                x = bytearray(x)
+                x = filter(lambda x : chr(x) in string.printable, x)
+                x = "".join([chr(c) for c in x])
             output += x
             x = x.rstrip()
             if show_output:
@@ -100,7 +114,7 @@ class build_binaries(object):
                           "--recursive",
                           "-f"])
 
-    def checkout(self, vehicle, ctag, cboard=None, cframe=None):
+    def checkout(self, vehicle, ctag, cboard=None, cframe=None, submodule_update=True):
         '''attempt to check out a git tree.  Various permutations are
 attempted based on ctag - for examplle, if the board is avr and ctag
 is bob we will attempt to checkout bob-AVR'''
@@ -114,7 +128,11 @@ is bob we will attempt to checkout bob-AVR'''
         if ctag == "latest":
             vtag = "master"
         else:
-            vtag = "%s-%s" % (vehicle, ctag)
+            tagvehicle = vehicle
+            if tagvehicle == "Rover":
+                # FIXME: Rover tags in git still named APMrover2 :-(
+                tagvehicle = "APMrover2"
+            vtag = "%s-%s" % (tagvehicle, ctag)
 
         branches = []
         if cframe is not None:
@@ -131,12 +149,12 @@ is bob we will attempt to checkout bob-AVR'''
             try:
                 self.progress("Trying branch %s" % branch)
                 self.run_git(["checkout", "-f", branch])
-                self.run_git_update_submodules()
+                if submodule_update:
+                    self.run_git_update_submodules()
                 self.run_git(["log", "-1"])
                 return True
             except subprocess.CalledProcessError:
                 self.progress("Checkout branch %s failed" % branch)
-                pass
 
         self.progress("Failed to find tag for %s %s %s %s" %
                       (vehicle, ctag, cboard, cframe))
@@ -172,7 +190,7 @@ is bob we will attempt to checkout bob-AVR'''
     def skip_frame(self, board, frame):
         '''returns true if this board/frame combination should not be built'''
         if frame == "heli":
-            if board in ["bebop", "aerofc-v1", "skyviper-v2450", "CubeSolo", "CubeGreen-solo"]:
+            if board in ["bebop", "aerofc-v1", "skyviper-v2450", "CubeSolo", "CubeGreen-solo", 'skyviper-journey']:
                 self.progress("Skipping heli build for %s" % board)
                 return True
         return False
@@ -224,12 +242,18 @@ is bob we will attempt to checkout bob-AVR'''
         with open(filepath, "w") as x:
             x.write(string)
 
+    def version_h_path(self, src):
+        '''return path to version.h'''
+        if src == 'AP_Periph':
+            return os.path.join('Tools', src, "version.h")
+        return os.path.join(src, "version.h")
+
     def addfwversion_gitversion(self, destdir, src):
         # create git-version.txt:
         gitlog = self.run_git(["log", "-1"])
         gitversion_filepath = os.path.join(destdir, "git-version.txt")
         gitversion_content = gitlog
-        versionfile = os.path.join(src, "version.h")
+        versionfile = self.version_h_path(src)
         if os.path.exists(versionfile):
             content = self.read_string_from_filepath(versionfile)
             match = re.search('define.THISFIRMWARE "([^"]+)"', content)
@@ -246,7 +270,7 @@ is bob we will attempt to checkout bob-AVR'''
 
     def addfwversion_firmwareversiontxt(self, destdir, src):
         # create firmware-version.txt
-        versionfile = os.path.join(src, "version.h")
+        versionfile = self.version_h_path(src)
         if not os.path.exists(versionfile):
             self.progress("%s does not exist" % (versionfile,))
             return
@@ -325,8 +349,13 @@ is bob we will attempt to checkout bob-AVR'''
         self.progress("Building %s %s binaries (cwd=%s)" %
                       (vehicle, tag, os.getcwd()))
 
-        for board in boards:
-            self.progress("Building board: %s" % board)
+        board_count = len(boards)
+        count = 0
+        for board in sorted(boards, key=str.lower):
+            now = datetime.datetime.now()
+            count += 1
+            self.progress("[%u/%u] Building board: %s at %s" %
+                          (count, board_count, board, str(now)))
             for frame in frames:
                 if frame is not None:
                     self.progress("Considering frame %s for board %s" %
@@ -335,14 +364,13 @@ is bob we will attempt to checkout bob-AVR'''
                     framesuffix = ""
                 else:
                     framesuffix = "-%s" % frame
-                if not self.checkout(vehicle, tag, board, frame):
+                if not self.checkout(vehicle, tag, board, frame, submodule_update=False):
                     msg = ("Failed checkout of %s %s %s %s" %
                            (vehicle, board, tag, frame,))
                     self.progress(msg)
                     self.error_strings.append(msg)
                     continue
-                if self.skip_board_waf(board):
-                    continue
+
                 self.progress("Building %s %s %s binaries %s" %
                               (vehicle, tag, board, frame))
                 ddir = os.path.join(self.binaries,
@@ -355,10 +383,21 @@ is bob we will attempt to checkout bob-AVR'''
                 if self.skip_frame(board, frame):
                     continue
 
+                # we do the submodule update after the skip_board_waf check to avoid doing it on
+                # builds we will not be running
+                self.run_git_update_submodules()
+
+                if self.skip_board_waf(board):
+                    continue
+                
                 if os.path.exists(self.buildroot):
                     shutil.rmtree(self.buildroot)
 
                 self.remove_tmpdir()
+
+                githash = self.run_git(["rev-parse", "HEAD"]).rstrip()
+
+                t0 = time.time()
 
                 self.progress("Configuring for %s in %s" %
                               (board, self.buildroot))
@@ -381,19 +420,34 @@ is bob we will attempt to checkout bob-AVR'''
                            (vehicle, board, framesuffix, tag))
                     self.progress(msg)
                     self.error_strings.append(msg)
+                    # record some history about this build
+                    t1 = time.time()
+                    time_taken_to_build = t1-t0
+                    self.history.record_build(githash, tag, vehicle, board, frame, None, t0, time_taken_to_build)
                     continue
+
+                t1 = time.time()
+                time_taken_to_build = t1-t0
+                self.progress("Building %s %s %s %s took %u seconds" %
+                              (vehicle, tag, board, frame, time_taken_to_build))
 
                 bare_path = os.path.join(self.buildroot,
                                          board,
                                          "bin",
                                          "".join([binaryname, framesuffix]))
                 files_to_copy = []
-                for extension in [".px4", ".apj", ".abin", "_with_bl.hex", ".hex"]:
+                extensions = [".px4", ".apj", ".abin", "_with_bl.hex", ".hex"]
+                if vehicle == 'AP_Periph':
+                    # need bin file for uavcan-gui-tool and MissionPlanner
+                    extensions.append('.bin')
+                for extension in extensions:
                     filepath = "".join([bare_path, extension])
                     if os.path.exists(filepath):
                         files_to_copy.append(filepath)
+                if not os.path.exists(bare_path):
+                    raise Exception("No elf file?!")
                 # only copy the elf if we don't have other files to copy
-                if os.path.exists(bare_path) and len(files_to_copy) == 0:
+                if len(files_to_copy) == 0:
                     files_to_copy.append(bare_path)
 
                 for path in files_to_copy:
@@ -404,6 +458,9 @@ is bob we will attempt to checkout bob-AVR'''
                 # why is touching this important? -pb20170816
                 self.touch_filepath(os.path.join(self.binaries,
                                                  vehicle_binaries_subdir, tag))
+
+                # record some history about this build
+                self.history.record_build(githash, tag, vehicle, board, frame, bare_path, t0, time_taken_to_build)
 
         if not self.checkout(vehicle, tag, "PX4", None):
             self.checkout(vehicle, "latest")
@@ -500,10 +557,8 @@ is bob we will attempt to checkout bob-AVR'''
 
     def common_boards(self):
         '''returns list of boards common to all vehicles'''
-        # note that while we do not use these for AntennaTracker!
         return ["fmuv2",
                 "fmuv3",
-                "fmuv4",
                 "fmuv5",
                 "mindpx-v2",
                 "erlebrain2",
@@ -514,11 +569,15 @@ is bob we will attempt to checkout bob-AVR'''
                 "pxfmini",
                 "KakuteF4",
                 "KakuteF7",
+                "KakuteF7Mini",
                 "MatekF405",
                 "MatekF405-STD",
                 "MatekF405-Wing",
+                "MatekF765-Wing",
+                "MatekH743",
                 "OMNIBUSF7V2",
                 "sparky2",
+                "omnibusf4",
                 "omnibusf4pro",
                 "omnibusf4v6",
                 "OmnibusNanoV6",
@@ -526,18 +585,27 @@ is bob we will attempt to checkout bob-AVR'''
                 "airbotf4",
                 "revo-mini",
                 "CubeBlack",
+                "CubeBlack+",
                 "CubePurple",
                 "Pixhawk1",
+                "Pixhawk1-1M",
                 "Pixhawk4",
+                "Pix32v5",
                 "PH4-mini",
                 "CUAVv5",
                 "CUAVv5Nano",
+                "CUAV-Nora",
+                "CUAV-X7",
                 "mRoX21",
                 "Pixracer",
                 "F4BY",
                 "mRoX21-777",
+                "mRoControlZeroF7",
+                "mRoNexus",
+                "mRoPixracerPro",
                 "F35Lightning",
                 "speedybeef4",
+                "SuccexF4",
                 "DrotekP3Pro",
                 "VRBrain-v51",
                 "VRBrain-v52",
@@ -545,18 +613,37 @@ is bob we will attempt to checkout bob-AVR'''
                 "VRCore-v10",
                 "VRBrain-v54",
                 "TBS-Colibri-F7",
-                "Pixhawk6",
+                "Durandal",
                 "CubeOrange",
                 "CubeYellow",
+                "R9Pilot",
+                "QioTekZealotF427",
                 # SITL targets
                 "SITL_x86_64_linux_gnu",
                 "SITL_arm_linux_gnueabihf",
                 ]
 
+    def AP_Periph_boards(self):
+        '''returns list of boards for AP_Periph'''
+        return ["f103-GPS",
+                "f103-ADSB",
+                "f103-RangeFinder",
+                "f303-GPS",
+                "f303-Universal",
+                "f303-M10025",
+                "f303-M10070",
+                "f303-MatekGPS",
+                "f103-Airspeed",
+                "CUAV_GPS",
+                "ZubaxGNSS",
+                "CubeOrange-periph",
+                "CubeBlack-periph",
+                ]
+
     def build_arducopter(self, tag):
         '''build Copter binaries'''
         boards = []
-        boards.extend(["skyviper-v2450", "aerofc-v1", "bebop", "CubeSolo", "CubeGreen-solo"])
+        boards.extend(["skyviper-v2450", "aerofc-v1", "bebop", "CubeSolo", "CubeGreen-solo", "skyviper-journey"])
         boards.extend(self.common_boards()[:])
         self.build_vehicle(tag,
                            "ArduCopter",
@@ -591,11 +678,11 @@ is bob we will attempt to checkout bob-AVR'''
         '''build Rover binaries'''
         boards = self.common_boards()
         self.build_vehicle(tag,
-                           "APMrover2",
+                           "Rover",
                            boards,
                            "Rover",
                            "ardurover",
-                           "APMrover2")
+                           "Rover")
 
     def build_ardusub(self, tag):
         '''build Sub binaries'''
@@ -606,10 +693,21 @@ is bob we will attempt to checkout bob-AVR'''
                            "ardusub",
                            "ArduSub")
 
+    def build_AP_Periph(self, tag):
+        '''build AP_Periph binaries'''
+        boards = self.AP_Periph_boards()
+        self.build_vehicle(tag,
+                           "AP_Periph",
+                           boards,
+                           "AP_Periph",
+                           "AP_Periph",
+                           "AP_Periph")
+
+        
     def generate_manifest(self):
         '''generate manigest files for GCS to download'''
         self.progress("Generating manifest")
-        base_url = 'http://firmware.ardupilot.org'
+        base_url = 'https://firmware.ardupilot.org'
         generator = generate_manifest.ManifestGenerator(self.binaries,
                                                         base_url)
         content = generator.json()
@@ -620,12 +718,19 @@ is bob we will attempt to checkout bob-AVR'''
         new_json_filepath_gz = os.path.join(self.binaries,
                                             "manifest.json.gz.new")
         with gzip.open(new_json_filepath_gz, 'wb') as gf:
+            if running_python3:
+                content = bytes(content, 'ascii')
             gf.write(content)
         json_filepath = os.path.join(self.binaries, "manifest.json")
         json_filepath_gz = os.path.join(self.binaries, "manifest.json.gz")
         shutil.move(new_json_filepath, json_filepath)
         shutil.move(new_json_filepath_gz, json_filepath_gz)
         self.progress("Manifest generation successful")
+
+        self.progress("Generating stable releases")
+        gen_stable.make_all_stable(self.binaries)
+        self.progress("Generate stable releases done")
+
 
     def validate(self):
         '''run pre-run validation checks'''
@@ -652,6 +757,10 @@ is bob we will attempt to checkout bob-AVR'''
             self.progress("Removing (%s)" % (self.tmpdir,))
             shutil.rmtree(self.tmpdir)
 
+    def buildlogs_dirpath(self):
+        return os.getenv("BUILDLOGS",
+                         os.path.join(os.getcwd(), "..", "buildlogs"))
+
     def run(self):
         self.validate()
 
@@ -660,7 +769,10 @@ is bob we will attempt to checkout bob-AVR'''
         origin_env_path = os.environ.get("PATH")
         os.environ["PATH"] = ':'.join([prefix_bin_dirpath, origin_env_path,
                                        "/bin", "/usr/bin"])
-        self.tmpdir = os.path.join(os.getcwd(), 'build.tmp.binaries')
+        if 'BUILD_BINARIES_PATH' in os.environ:
+            self.tmpdir = os.environ['BUILD_BINARIES_PATH']
+        else:
+            self.tmpdir = os.path.join(os.getcwd(), 'build.tmp.binaries')
         os.environ["TMPDIR"] = self.tmpdir
 
         print(self.tmpdir)
@@ -682,8 +794,7 @@ is bob we will attempt to checkout bob-AVR'''
 
         self.mkpath(os.path.join("binaries", self.hdate_ym,
                                  self.hdate_ymdhm))
-        self.binaries = os.path.join(os.getcwd(), "..", "buildlogs",
-                                     "binaries")
+        self.binaries = os.path.join(self.buildlogs_dirpath(), "binaries")
         self.basedir = os.getcwd()
         self.error_strings = []
 
@@ -697,11 +808,14 @@ is bob we will attempt to checkout bob-AVR'''
                                       "binaries.build")
 
         for tag in self.tags:
+            t0 = time.time()
             self.build_arducopter(tag)
             self.build_arduplane(tag)
             self.build_rover(tag)
             self.build_antennatracker(tag)
             self.build_ardusub(tag)
+            self.build_AP_Periph(tag)
+            self.history.record_run(githash, tag, t0, time.time()-t0)
 
         if os.path.exists(self.tmpdir):
             shutil.rmtree(self.tmpdir)
